Add article subscriptions.
The idea is to allow users to subscribe to articles in order for them
to receive updates (via email for example) when changes are made
to the articles they subscribe to.

The purpose is to avoid stale information not just in the
documentation database itself, but inside the minds of the team
as well.