Ensure local images can't be overridden by OAuth.
If someone uploaded an image manually, the image was replaced
as soon as the user in question had to log back into Orientation
because the OAuth authentication process always caused the
associated user image to be updated.

This is no longer the case and the OAuth image will only be used
for new users who never had an image before.