Add ability to make articles rotten or refresh them.
This contains a suite of changes. First I refactored the #make_fresh
controller action that was doing direct DB manipulation on the article
by calling touch(:updated_at), and instead wrapped that behavior
inside of a model instance method (#refresh!).

I added specs for that in all three possible states: fresh, stale
and rotten.

Which brings me to the new state: rotten.

I added the a similar introspection method called #rotten? which just
like #fresh? delegates to a class method. Yes I was being overly
fancy when I did that but I'd rather follow the pattern for now.

An article can now be marked as rotten. In essence it means the article
is now obsolete and requires its author or editor's attention. Later on
we should add email notification logic that will ping both the author
and the editor (if it exists).

For now there is a controller action called #make_rotten but I haven't
included a "Mark as Rotten" button in the article show view.