Clean up fresh/stale/rotten interfaces.
I don't think it made sense to direct authors to the edit action when linking
them to an article from the staleness mailer. Authors needs to read the article
to even remember what it's about first because they can gauge whether
it's stale or not.

If it is stale, then they can edit it and they will be take back to show where
they'll have the option to "Mark Fresh" the article. Ditto if it's rotten.

If the article isn't rotten yet, we offer a button to mark it as such, whether it was
already flagged as stale automatically or not.

I also cleaned up the hardcoded HTML in the staleness notification mailer,
content_tag and ActionView::Helpers::UrlHelper are made for that.