Trust Draper to do its job again
This is kind of harsh. I suspect that Draper wasn't able to forward the 
call to #primary_key on `@article.subscribe(current_user)` because 
despite AuthorDecorator inheriting from UserDecorator it didn't know 
that it was supposed to decorate a User class and therefore find the
primary key on the User object.

This is a shot in the dark, I haven't tested this.