Add Article#old_versions and #has_old_versions?
Also fix a weird bug with original versions on the versions contoller so
that we stop showing the original version (which has no object) and
properly diff between the actual content (thankfully saved in the
object_changes column) when comparing the second version with the
original.

The paper_trail folks are being original and decided that when you
create something, the original state of the object doesn't matter.
That's fucking stupid and we'll perhaps patch that ridiculous behavior
later on.