Allow articles to be turned into guides.
Guides are simply articles which are used as guides to a
selection of other articles. Any article can be turned into a
guide and it will appear in the Guides section with its content
displayed entirely there.

This makes it very simple to curate a list or a simple short piece of
text to be used as a guide.