Attempt to fix erroneous exit 0 when specs fail
Unix programs exit with the code 0 when everything's cool. Otherwise
they exit with 1 to signal an exception. Due to some weird bug in
simplecov, (apparently before 0.7.1) the error code from rspec is
swallowed when simplecov runs and returns its own code. That's
mega-dumb and hopefully this update solves this.

Prior to this, failings specs were marked green on Travis because
Travis trusts the last exist code.