Fix ForbiddenAttributesError on sessions#create
Fix #96. Request parameters were being passed directly to an external
class and that's somehow not okay anymore since 4.1.x.

Fixed by calling #to_h on the request param sub-hash we do care about.
I don't see an obvious security issue with this but I need to write a
spec a properly authorize the relevant parameters in order to not
circumvent strong params anymore.