Only send article updates on 5 minute delay.
This way, if another update is made within the five
minutes, the previous delayed job is deleted and
replaced with a new one for the most recent update.

That should solve the problem of people editing
articles repeatedly and triggering a rain of emails
on subscribers.

Also includes a slight refactor of ArticleSubscription#send_update
which did not need to be passed an article ID as it holds a reference
to one on each instance.