Fix missing counter_cache on associations
I setup a counter_cache on Article for endorsements
and subscriptions but forgot to specify the cache
on each individual association as well.

Rails uses the current class name to infer the counter
cache column on the associated model (Article) so
here I had to manually declare the name.

It's for the best because I'd rather not have to call
article.article_subscriptions on the other end.