Add Slack notifications for article updates
This is a furiously rushed feature, once again blame is
solely on @katiedelfin for whispering this fabulous
idea in my ear at the best possible moment.

I expect to eventually spec out Speakerphone and
clean it up a bit if need be. Ditto for some of the
methods added to Article. I'm actually surprised AR
doesn't have an updated_at/created_at comparison.
It seems like such a simple and elegant way to check
for new record vs updated record.

Obviously as you can tell by the half-assed
dependency injection of the audience argument in
Speakerphone, the idea is to let pepple shout to
whatever they prefer, not just Slack.

But Slack's the bomb diggity so if you want another
integration, you gotta work that PR, my friend. ;-)