Separate articles & guides on Guides index
The point is that when searching we don't want to
limit things to only guides, it's making it seem like
there are no articles beyond guides.

The guides index pre-search is a useful starting point
for discovering through browsing, but Article keyword
searching should work just like it does on the Article
index.