Extract article search into partial
Very important detail is that the form action URL
is now based on the current scope. This means that
search is no longer broken on the Rotten articles
page. It returns only results from the rotten scope and
not articles from every scope.

This is great!

We also don't have duplication in the JS search DOM
replacement handling anymore. And I'm pretty sure
that's going to make @johndjameson happy. :-)