Pass current scope to Article.text_search
This allows the search method to always return all
values from the current scope as opposed to all
values period (the list of all articles).

Important detail I didn't mention in the commit title:
we're now back using PostgreSQL full-text fuzzy
search but only on article titles. It seems like
something's fucked up with the indexing on the
content column which makes fuzzy searching on it
not work properly. Besides in order for fuzzy search to
make sense from the end-user perspective we'd have
to find a way to give them a preview of the section of
content that matched otherwise it's very hard to tell
which match contains what they're looking for since
we only show them the article titles in search results.