Redirect to articles#new on 404
Grab the slug and titleize it to extract a potential title and
encourage people to write articles that don't already exists.

This makes sense now that it's unlikely that 404 will happen due to
oudated slugs since we retain slug history and redirection of old
slugs to new ones through FriendlyId.