Create Teams
Users can now be associated with a team.

The team can have a name and a shtick just like normal authors but
nothing else to keep things simple.

Eventually I'd like to have teams be displayed the same way Guides are
currently, as a mosaic of blocks with the name of the team, the shtick
and small circle avatars of each team member up to a reasonable limit.

Curious what @drewbarontini and @johndjameson will think of this one. :-)