Clarify the purpose of Guides
I noticed that people were misunderstanding "Guide" to mean "Tutorial"
and checking the "Guide" box on the Article form because the article
was a "How-To Guide" essentially.

They didn't seem to notice that Guides are made to be very high-level since
they're displayed on the home page so this little prose change hopefully
will partially address that.