Create ArticlesTag model to more easily deal with the join table 'articles_tags'
This model will allow us to clean up counter cache code in a following commit,
as well as provide a case-insensitive uniqueness validation to prevent us from
adding the same tag to an article multiple times. There is a migration to add
a primary key to the join table, in order for Rails to be able to do certain
operations (like the uniqueness validation).