Mark invalid internal links
This means it's now possible to create a link to a non-existent
article and be able to see that it doesn't match an existing article.

This should allow folks to create lists of articles to create and tackle
them one by one by making the red links go blue. Should be fun. :-)