Switch from fuzzy to advanced search on article titles
This has been a persistent issue even after improving fuzzy_search a bit recently.

The expectation is simple: people look for article titles; they rarely search for things that they
expect within article content. Regardless, the two searches collide and create huge problems
for us for something as simple as an article titled: iOS App - Testing with TestFlight

With fuzzy_search on title and content:

- ios: no match
- ios app: no match
- ios app testing: match
- testing: no match
- testing with: match
- testing with testflight: match
- iOS testflight: match
- iOS testflight testing: match
- app testing: match
- app test: no match
- with iOS: match
- ios test: no match

Whereas witih avanced_search on title alone:
- ios: match
- ios app: match
- ios app testing: match
- testing: match
- testing with: match
- testing with testflight: match
- iOS testflight: match
- iOS testflight testing: match
- app testing: match
- app test: match
- with iOS: match
- ios test: match

To me the results are pretty clear. Of course this isn't to say there isn't a better way
(tokenized search) to do title AND content searching but so far I haven't figured out how to
do that properly.