Log rot reporter & mention them in notification
Now when someone reports an article as rotten we log who it was.

Why?
- it makes things more human, it's not just a game of whack-a-mole to
find articles and mark them as rotten
- it allows for conversations to happen between the reporter and the
article contributors
- if notifications are turned off it allows a reader or contributor of the
article to reach out to the rot reporter in order to refresh the article

This commit also refactors the ArticleRottenJob to put all the fetching
logic inside of it. We don't pass a contributors object, or an article, or
a reporter. All of those are fetched at job runtime.

I'll have to update the Mandrill templates in order to use the now
added REPORTER_NAME and REPORTER_URL merge vars.

Fixes #178