Add CUSTOM_RUBY_VERSION and default version
Once again following advice from:
https://devcenter.heroku.com/articles/ruby-versions#specifying-a-ruby-version-via-the-environment

We didn't specify the Ruby version in the Gemfile before but this has
become a problem since Heroku will only read it from the Gemfile.

We now tell Travis to override this default with the CUSTOM_RUBY_VERSION
in the environment variable for the Ruby 2.1 build. See .travis.yml.