Revert "Adds Emoji support"
This reverts commit ad121fdd94907801f68424eff83e55c3b63d8314.

This fixes a rather serious issue raised in #206. I don't
think the value of emoji is worth the bugs introduced so
I feel bad for @BrandonMathis because it's a fun feature
so if there's anyway to re-introduce it in a way that does
not break something else, I'm game.

In the meantime, we'll do without it, I'm afraid.

# Conflicts:
#	Gemfile.lock
#	app/helpers/application_helper.rb