Actually add the Markdown image tag to editor
This is still hacky in that I specifically remove the part of the absolute path that is specific
to the local machine. There should be a way to obtain just the relative path from the public
folder from the Rails controller and only send that back but this works.

Now I have to figure out the whole remote upload to S3 sort of deal which I have a feeling
will take much longer.