Configure Orientation via orientation.yml
This is a prototype with a request for comments. I think it's time for Orientation
to move away from "everybody has to use everything" model and let people
disable features in a sane way. Or better yet, let them use alternatives.

We use Mandrill but you might use SendGrid.
We use Google for OAuth but you might have your own OAuth server.
We use Bugsnag & Skylight but you might not.
Etc.