Article versioning display fixes
This commit fixes some issues with article versioning display and adds a
title only search. The postgres full text search occasionally produces
unexpected results when searching for, for example, a title of "About".
The article exists but is not returned. Therefore, a "title only" text
search was added as a work around.