Create article_views table
This will be used to store the first and last time a user viewed an article.

Used as an opportunity to replace structure.sql with schema.rb

Rails finally supports complex indices like tsvector so we don't need
to use the atrociously ugly structure.sql to ensure that those are
properly created on db:schema:load.

Signed-off-by: Olivier Lacan <hi@olivierlacan.com>