Add unique index on article_views
We don't want to create a new view every single time a user views an
article. Instead we simply want to update the existing view if one
already exists.

This allows us to know when a user viewed an article for the first time
(create_at), and the last time a user viewed an article (updated_at).