Add Article::View model
Note the require_dependency trick here.

I could have done:

    class Article < ActiveRecord::Base
       class View < ActiveRecord::Base
       end
    end

But I hate reopening classes simply to define a namespace
under an existing model class. Defining this class as
Article::View will do a constant lookup on Article first. If for
any reason Rails hasn't autoloaded the existing Article class
first, it may try to define an Article module to make the
namespacing work for the Article::View class.

`require_dependency "article"` forces Rails to check for an
existing Article class before defining this class, which solves
the constant lookup issue in a single line and clearly establishes
the relationship between the two classes, which is a bonus in
my book.