Add count integer column to article_views table
We already store anonymous article visits through the articles.visits
column. This, instead, allows us to know how many times a specific
visitor (since Article::View is associated to a User) has visited a specific
article.

This established a better picture of the relationship between a user and
specific articles that goes beyond the manual decision to subscribe to
an article.

From experience, few people manually decide to subscribe but they do
regularly come back to the same piece of content. Not being subscribed
makes it more difficult for them to notice what has changed since the last
time they visited.

Since we store the user's last visit date in the article_views.updated_at column
we could do the following:
- tell user what has been updated since their last visit (this would require some
sort of diffing mechanism which is being worked on in the "versioning" branch)
- suggest that users who have a high articles_views.count for a specific article
should subscribe to the article
- depending on a user's specific team, we could map out articles which are
often viewed by a specific team when a new user joins a team

I'm sure we can find other interesting ways to use this information in the future.

What's nice however is that since we do collect anonymous visits on the articles
table, we don't need to depend on article_views (which may be disabled) to provide
a popularity metric to the full-text search algorithm, so people who have a privacy
concern should be able to easily disable user-specific metrics gathering.