Rename "rotten" articles to "outdated"
There is such a thing as pushing a metaphor too far. I've always
felt a little uneasy about my choice of "rotten" to describe articles
that are wholly untrustworthy.

I don't mind the colorful and memorable image, it's more the value-judgement
associated with it that I believe could prevent people from marking an article
written by someone else as "rotten" because that sounds fucking harsh.

While "outdated" is definitely more clinical and boring, it's also a judgement
on the state of the article, not on the quality of its content.