Refactor article show view
Extracted a ton of article-related things from this view into partials. This
will cause a performance hit but I expect to migrate to hamlit soon to negate
that performance issue.

Better yet, renamed the "Fresh" and "Outdated" buttons to "Mark Fresh" and
"Mark Outdated" to clarify their purpose. And to top it off I've also added
tooltips that further explain what the toggles are intended for.

# Conflicts:
#	app/views/articles/show.html.haml