Remove untestable example
Seems like Capybara and rack-test don't provide a way to
find URL fragments.

I could have tried to find a change in the viewport but that seems
far-fetched too. I'll try to revisit this later to ensure that
the header anchors are not broken once again in the future
without anyone noticing.