Remove the app layout when responding to XHR
This is a very strange issue. There was previously no need to
remove the layout (although it probably should have been
disabled) but suddenly the self-calling function inside index.js.erb
stopped executing on the client-side and I couldn’t figure out
when until I realized that it was sending the whole application
layout and therefore preventing the JS from executing.