Double the search input delay to prevent races
The full story is in #247. Essentially fast typists will type multiple
letters into the search input field in less than half a second.

Since our input delay is 200 ms (.2 second) then it’s possible for
someone to type one letter every .2 second which will fire one
XHR search query to the server for each and every letter.

This is responsive but not ideal because of the asynchronous
nate of these queries. It’s very possible that when searching for
“monday”, the response for “mon” will arrive before the
response for “m”. And when “m” does arrive, it will wipe out the
search results for “mon” (which are probably good) with a useless
result since single-letter queries can’t possibly yield a full-text
match.

Another way around this issue would be to not send any request
to the server until we have three letters in the search input field.

I’m pretty sure I used to have it set up that way too.