Replace input delay with query length check
Now instead of delaying all search calls we simply ignore
what’s being typed in the search input field until it’s 3 characters
at least.

But we have to let 0 characters fire a call so that the list of all
articles can be displayed when people remove their search query.