Hide inactive users from endorsers & subscribers
People leave, and when they do it's always odd to see a broken Google
avatar of them endorsing an article.

This shouldn't affect the endorsement and subscription counts. So an
article can still benefit from old endorsements and subscriptions to
increase its perceived value. Although this may not make sense in the
long term, since by definition inactive users aren't there to see if
they should still be endorsing an article.