Replace DelayedJob Jobs with Sidekiq Workers
In the process, I reinstated the staleness notification for articles which will
need to be tuned to send when an article becomes stale to avoid a barrage of
emails to authors.

I replaced the handcrafted 5 minute delay and unique article updated notification
check with the sidekiq-unique-jobs library which does the same thing with a
TON more code. So I might replace that at some point later when I’m more
familiar with the Sidekiq API because it doesn’t seem like it should that this
much code to simply avoid duplicate jobs in a 5 minute window.

I had to update database_cleaner which was relying on the deprecated Fixnum
class (Ruby 2.4) as well. Ditto with Redis.

Finally I added some proxy methods for Article Subscription and Endorsement
so that the user associated with both can be referred to as subscriber or
endorser for clarity.

Oh, and User#inactive? is often a lot more useful than User#active?. Gotta
replace that thing with a timestamp column to remember when users where
marked as inactive. Not a biggie but it could be useful.

Yes, this commit has too much stuff. Sue me.