Finally fix Article.text_search ordering
Years to fix this damn thing and now we finally get proper prioritization
because all it took was removing the order from the existing scope because
all that really matters on a search is the order that the full-text search engine
gave.

This is the worst and best thing all wrapped into one.