Remove redundant default ports
If you use docker-compose the db container should be available
with the `db` hostname and the default postgres port. There
should be no need to specify port mapping to the host machine
because with this setup Docker also runs Postgres and there’s no
communication with the host machine other than exposing port 3000
for localhost:3000 to be accessible in the browser.