Replace Pygments.rb with Rouge
Long time coming but now that Pygments.rb is barely maintained, that it depends
on a very old version of Python (2.7) and complicates the installation process
for Orientation, it was about time to switch to the most excellent all-Ruby
alternative: http://rouge.jneen.net

I can't believe how easy it was to switch over. Rouge even aliases `shell`
to `bash` and vice-versa which makes the lexer hack I had with Pygments
unnecessary (it would blow up on languages it didn't know).

Rouge on the contrary just doesn't do any highlighting for any language passed
in the Markdown fenced codeblocks that it can't recognize (triple backticks
followed by the language name in lowercase)

What a relief!

Oh, and I took the opportunity to refresh the CSS we're using to display
highlighted code to switch to the GitHub styles which are quite appropriate for
the current Orientation theme.