Fix OmniAuth security vulnerability
This will eventually mean adding non-seamless step to authenticate to Google
which makes a lot of sense (it requires a post request from a link) in general
so people aren't redirected to an OAuth provider without warning.

I'll have to design a new page for that and it will actually make customizing
the OAuth provider that much easier.